controllers

+++ src/main/java/org/example/models/User.java
@@ -1,9 +1,6 @@
 package org.example.models;
 
 import javax.persistence.*;
-import javax.validation.constraints.Min;
-import javax.validation.constraints.NotEmpty;
-import javax.validation.constraints.Size;
 
 @Entity
 @Table(name = "user")
@@ -11,7 +8,7 @@ public class User {
     @Id
     @Column(name = "id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
 //    @NotEmpty(message = "Name should not be empty")
 //    @Size(min = 2, max = 30, message = "Name should be between 2 and 30 characters")
@@ -29,11 +26,11 @@ public class User {
     public User() {
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
